Changed the pilot ID type to uint32_t.
This will make the binary search done to identify a pilot faster if
I understand correctly, and 32-bit is plenty to avoid wraparound
problems.

+++ src/ai.h
@@ -12,7 +12,7 @@
 #include "physics.h"
 
 /* Forward declaration to avoid cyclical import. */
-typedef unsigned long pilotId_t;
+typedef uint32_t pilotId_t;
 struct Pilot_;
 typedef struct Pilot_ Pilot;
